+++ test_settings.py
@@ -2,6 +2,9 @@ from __future__ import annotations
 
 import os
 
+from django.urls import include
+from django.urls import path
+
 PROJECT_ROOT = os.path.normpath(os.path.join(os.path.dirname(__file__), "payments"))
 TEMPLATES = [
     {
@@ -14,3 +17,16 @@ SECRET_KEY = "NOTREALLY"
 PAYMENT_HOST = "example.com"
 
 INSTALLED_APPS = ["payments", "django.contrib.sites"]
+
+ROOT_URLCONF = "test_settings"
+
+urlpatterns = [
+    path("payments/", include("payments.urls")),
+]
+
+DATABASES = {
+    "default": {
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": ":memory:",
+    }
+}